refactor(provider): isolate chat template fixes by model
